Case 4: Test Case for IEEE 14 Bus Power Flows.
The IEEE 14-bus test case represents a portion of the US electrical power system (located in the US Midwest), the information presented here is dated February 1962. A copy of the data was kindly provided. provided by Iraj Dabbagchi of AEP (American Electric Power) and introduced to the IEEE common format by Rich Christie of the University of Washington in 1993. The system is made up of fourteen buses, five generators, five transformers, sixteen lines and eleven loads. The voltage levels on the bars and the power limits on the lines and transformers were taken from the book Simulation of Electrical Systems.