Case 5: Test Case for IEEE 30 Bus Power Flows.
The IEEE 30-bus test case represents a portion of the US electric power system (located in the US Midwest). In December 1961. The data was provided by Iraj Dabbagchi AEP and entered into the IEEE common format by Rich Christie at the University of Washington in August 1993. The impedance data is referred to a base power of 100 MVA.